fix CV making circular imports
